+++ CHANGELOG.md
@@ -6,6 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2024-05-17
+### Changed
+- `HasError` trait was rejigged to be more flexible.
+  - Supports a suffix (in addition to the prefix it already did.)
+  - Supports using colours.
+- `ParamsApp` includes new `HasMessages` trait.
+### Added
+- A new `HasMessages` trait offering different types of status messages.
+  - Supports the same kinds of formatting options as `HasError`.
+  - Has a generic `msg()` method for adding additional message types.
+- A new `Themes/Standard` trait to colourize status messages and errors.
+
 ## [1.1.0] - 2024-01-11
 ### Added
 - New `Util` static class with some helper functions.
@@ -19,7 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pulls all `Lum\CLI` libraries out of `lum-framework`.
 - Also pulls `Fakeserver` library out of `lum-core`.
 
-[Unreleased]: https://github.com/supernovus/lum.cli.php/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/supernovus/lum.cli.php/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/supernovus/lum.cli.php/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/supernovus/lum.cli.php/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/supernovus/lum.cli.php/releases/tag/v1.0.1